Improve sameLocationType feature.

+++ src/main/scala/edu/knowitall/repr/bestmention/ResolvedBestEntityMention.scala
@@ -29,7 +29,7 @@ trait Coref {
 
 trait CandidateCount {
   this: BestMention =>
-  def candidateCount: Int
+  def candidateCount: Double
 } 
 
 
@@ -40,31 +40,31 @@ case class IdentityBestMention(target: Entity) extends ResolvedBestMention {
 
 trait ResolvedBestMention extends BestMention with TargetResolved with CandidateCount
 object ResolvedBestMention {
-  case class ResolvedBestMentionImpl(target: Entity, bestMention: String, candidateCount: Int) extends ResolvedBestMention
-  def apply(target: Entity, bestMention: String, candidateCount: Int) = ResolvedBestMentionImpl(target, bestMention, candidateCount)
+  case class ResolvedBestMentionImpl(target: Entity, bestMention: String, candidateCount: Double) extends ResolvedBestMention
+  def apply(target: Entity, bestMention: String, candidateCount: Double) = ResolvedBestMentionImpl(target, bestMention, candidateCount)
 }
 
 trait FullResolvedBestMention extends ResolvedBestMention with BestResolved
 object FullResolvedBestMention {
-  case class FullResolvedBestMentionImpl(target: Entity, bestEntity: Entity, candidateCount: Int) extends FullResolvedBestMention
-  def apply(target: Entity, bestEntity: Entity, candidateCount: Int) = FullResolvedBestMentionImpl(target, bestEntity, candidateCount: Int)
+  case class FullResolvedBestMentionImpl(target: Entity, bestEntity: Entity, candidateCount: Double) extends FullResolvedBestMention
+  def apply(target: Entity, bestEntity: Entity, candidateCount: Double) = FullResolvedBestMentionImpl(target, bestEntity, candidateCount: Double)
 }
 
 trait CorefResolvedBestMention extends ResolvedBestMention with Coref
 object CorefResolvedBestMention {
-  case class CorefResolvedBestMentionImpl(target: Entity, bestMention: String, cluster: MentionCluster, candidateCount: Int) extends CorefResolvedBestMention
-  def apply(target: Entity, bestMention: String, cluster: MentionCluster, candidateCount: Int) = 
+  case class CorefResolvedBestMentionImpl(target: Entity, bestMention: String, cluster: MentionCluster, candidateCount: Double) extends CorefResolvedBestMention
+  def apply(target: Entity, bestMention: String, cluster: MentionCluster, candidateCount: Double) = 
     CorefResolvedBestMentionImpl(target, bestMention, cluster, candidateCount)
 } 
 
 trait CorefFullResolvedBestMention extends CorefResolvedBestMention with FullResolvedBestMention
 object CorefFullResolvedBestMention {
-  case class CorefFullResolvedBestMentionImpl(target: Entity, bestEntity: Entity, cluster: MentionCluster, candidateCount: Int) extends CorefFullResolvedBestMention
-  def apply(target: Entity, bestEntity: Entity, cluster: MentionCluster, candidateCount: Int) = 
+  case class CorefFullResolvedBestMentionImpl(target: Entity, bestEntity: Entity, cluster: MentionCluster, candidateCount: Double) extends CorefFullResolvedBestMention
+  def apply(target: Entity, bestEntity: Entity, cluster: MentionCluster, candidateCount: Double) = 
     CorefFullResolvedBestMentionImpl(target, bestEntity, cluster, candidateCount)
 } 
 
-case class LinkResolvedBestMention(m: Mention, link: FreeBaseLink, cluster: MentionCluster, candidateCount: Int) extends CorefFullResolvedBestMention {
+case class LinkResolvedBestMention(m: Mention, link: FreeBaseLink, cluster: MentionCluster, candidateCount: Double) extends CorefFullResolvedBestMention {
   import LinkResolvedBestMention._
   val bestEntity = linkEntity(link)
   val target = Entity(m.text, m.offset, m.text, bestEntity.entityType)
@@ -90,13 +90,13 @@ trait ContainerBestMention extends ResolvedBestMention {
   def bestMention = BestMentionFinderOriginalAlgorithm.locationCasing(text + ", " + containerEntity.name)
 }
 object ContainerBestMention {
-  case class ContainerBestMentionImpl(target: Entity, containerEntity: Entity, candidateCount: Int) extends ContainerBestMention
-  def apply(target: Entity, containerEntity: Entity, candidateCount: Int) =
+  case class ContainerBestMentionImpl(target: Entity, containerEntity: Entity, candidateCount: Double) extends ContainerBestMention
+  def apply(target: Entity, containerEntity: Entity, candidateCount: Double) =
     ContainerBestMentionImpl(target, containerEntity, candidateCount)
 }
 
 case class ContainmentBestMention(
-    target: Entity, containedEntity:Entity, containerEntity: Entity, candidateCount: Int) extends ContainerBestMention {
+    target: Entity, containedEntity:Entity, containerEntity: Entity, candidateCount: Double) extends ContainerBestMention {
   
   override def bestMention = containedEntity.name + ", " + containerEntity.name
 }
